+++ src/internal/layout-engine/engine-cache.ts
@@ -2,18 +2,18 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Position } from "../utils/position";
-import { LayoutEngineStepState } from "./engine-step";
+import { LayoutEngineState } from "./engine-state";
 
 export class LayoutEngineCacheNode {
   public position: null | Position = null;
-  public state: LayoutEngineStepState;
+  public state: LayoutEngineState;
   private next = new Array<LayoutEngineCacheNode>();
 
-  constructor(state: LayoutEngineStepState) {
+  constructor(state: LayoutEngineState) {
     this.state = state;
   }
 
-  matches(position: Position, compute: () => LayoutEngineStepState): LayoutEngineCacheNode {
+  matches(position: Position, compute: () => LayoutEngineState): LayoutEngineCacheNode {
     for (const nextNode of this.next) {
       if (nextNode.position!.x === position.x && nextNode.position!.y === position.y) {
         return nextNode;
