+++ src/internal/layout-engine/interfaces.ts
@@ -38,3 +38,8 @@ export interface LayoutShift {
   moves: readonly CommittedMove[];
   conflicts: readonly ItemId[];
 }
+
+export interface Conflicts {
+  items: ReadonlySet<ItemId>;
+  direction: Direction;
+}
